+++ CHANGELOG.md
@@ -34,6 +34,10 @@ Please visit [cucumber/CONTRIBUTING.md](https://github.com/cucumber/cucumber/blo
 
 ### Changed
 
+- In `DataTable#map_column`, Changed the `strict` argument into a keyword argument.
+  See [UPGRADING.md](./UPGRADING.md#upgrading-to-800).
+  ([Issue#1592](https://github.com/cucumber/cucumber-ruby/issues/1592))
+
 ### Removed
 
 - `AfterConfiguration` has been removed. Please use `InstallPlugin` or `BeforeAll` instead.